Remove unnecessary prop and CMR cleanup
